+++ node/include/cocaine/service/node/overseer.hpp
@@ -28,18 +28,6 @@ class id_t;
 }  // namespace cocaine
 
 namespace cocaine {
-namespace detail {
-namespace service {
-namespace node {
-
-class slave_t;
-
-}  // namespace node
-}  // namespace service
-}  // namespace detail
-}  // namespace cocaine
-
-namespace cocaine {
 namespace slave {
 struct channel_t;
 
@@ -64,81 +52,32 @@ class stream_t;
 
 namespace cocaine {
 
-using detail::service::node::slave_t;
+class engine_t;
 
-class overseer_t:
-    public std::enable_shared_from_this<overseer_t>
-{
-public:
-    enum class despawn_policy_t {
-        graceful,
-        force
-    };
-
-    typedef std::unordered_map<
-        std::string,
-        slave_t
-    > pool_type;
-
-    struct channel_wrapper_t {
-        slave::channel_t channel;
-        trace_t trace;
-
-        auto operator*() -> slave::channel_t& { return channel; }
-        auto operator*() const -> const slave::channel_t& { return channel; }
-        auto operator->() -> slave::channel_t* { return &channel; }
-        auto operator->() const -> const slave::channel_t* { return &channel; }
-    };
-
-    typedef std::deque<channel_wrapper_t> queue_type;
-
-private:
-    const std::unique_ptr<logging::logger_t> log;
-
-    context_t& context;
-
-    /// Time point, when the overseer was created.
-    const std::chrono::system_clock::time_point birthstamp;
-
-    /// The application manifest.
-    const manifest_t manifest_;
-
-    /// The application profile.
-    synchronized<profile_t> profile_;
-
-    /// IO loop for timers and standard output fetchers.
-    std::shared_ptr<asio::io_service> loop;
-
-    /// Slave pool.
-    synchronized<pool_type> pool;
-    std::atomic<int> pool_target;
-
-    /// Pending queue.
-    synchronized<queue_type> queue;
-
-    /// Statistics.
-    stats_t stats;
+class overseer_t {
+    std::shared_ptr<engine_t> engine;
 
 public:
     overseer_t(context_t& context, manifest_t manifest, profile_t profile,
         std::shared_ptr<asio::io_service> loop);
 
+    /// TODO: Docs.
     ~overseer_t();
 
-    /// Returns copy of the current manifest.
-    ///
-    /// Application's manifest is considered constant during all app's lifetime and can be
-    /// changed only through restarting.
-    auto manifest() const -> manifest_t;
+    /// Returns application total uptime in seconds.
+    auto uptime() const -> std::chrono::seconds;
 
-    /// Returns copy of the current profile, which is used to spawn new slaves.
+    /// Returns the copy of current profile, which is used to spawn new slaves.
     ///
     /// \note the current profile may change in any moment. Moveover some slaves can be in some kind
     /// of transition state, i.e. migrating from one profile to another.
     auto profile() const -> profile_t;
 
-    /// Returns application total uptime in seconds.
-    auto uptime() const -> std::chrono::seconds;
+    /// Returns the copy of current manifest.
+    ///
+    /// Application's manifest is considered constant during all app's lifetime and can be
+    /// changed only through restarting.
+    auto manifest() const -> manifest_t;
 
     /// Returns the complete info about how the application works using json-like object.
     auto info(io::node::info::flags_t flags) const -> dynamic_t::object_t;
@@ -158,10 +97,8 @@ public:
     /// \param id represents slave id to be enqueued (may be none, which means any slave).
     ///
     /// \todo consult with E. guys about deadline policy.
-    std::shared_ptr<client_rpc_dispatch_t>
-    enqueue(io::streaming_slot<io::app::enqueue>::upstream_type downstream,
-            app::event_t event,
-            boost::optional<slave::id_t> id);
+    auto enqueue(io::streaming_slot<io::app::enqueue>::upstream_type downstream, app::event_t event,
+                 boost::optional<slave::id_t> id) -> std::shared_ptr<client_rpc_dispatch_t>;
 
     /// Enqueues the new event into the most appropriate slave.
     ///
@@ -173,9 +110,8 @@ public:
     /// \param event an invocation event.
     /// \param id represents slave id to be enqueued (may be none, which means any slave).
     /// \return a tx stream.
-    auto enqueue(std::shared_ptr<api::stream_t> rx,
-        app::event_t event,
-        boost::optional<slave::id_t> id) -> std::shared_ptr<api::stream_t>;
+    auto enqueue(std::shared_ptr<api::stream_t> rx, app::event_t event,
+                 boost::optional<slave::id_t> id) -> std::shared_ptr<api::stream_t>;
 
     /// Tries to keep alive at least `count` workers no matter what.
     ///
@@ -196,40 +132,6 @@ public:
     /// \note after successful accepting the balancer will be notified about pool's changes.
     io::dispatch_ptr_t
     prototype();
-
-    /// Cancels all asynchronous pending operations, preparing for destruction.
-    void
-    cancel();
-
-private:
-    /// Spawns a new slave using current manifest and profile.
-    void
-    spawn(pool_type& pool);
-
-    /// \warning must be called under the pool lock.
-    void
-    assign(slave_t& slave, slave::channel_t& payload);
-
-    /// Seals the worker, preventing it from new requests.
-    ///
-    /// Then forces the slave to send terminate event. Starts the timer. On timeout or on response
-    /// erases slave.
-    void
-    despawn(const std::string& id, despawn_policy_t policy);
-
-    std::shared_ptr<control_t>
-    on_handshake(const std::string& id,
-                 std::shared_ptr<session_t> session,
-                 upstream<io::worker::control_tag>&& stream);
-
-    void
-    on_slave_death(const std::error_code& ec, std::string uuid);
-
-    void
-    rebalance_events();
-
-    void
-    rebalance_slaves();
 };
 
-}
+}  // namespace cocaine
